Move model component for target into components package
